CSS Overhaul and Separate JS File
CSS was remade to be more reusable on other pages and sections of the website. It also works slightly differently as well, allowing for landing pages that span the entire width of the page.
The embedded JavaScript test was also moved to an external file and linked into both pages. Arrows were added to blog page.